Includes a dummy command 43 to SD model. U-boot tries to execute this command even thought it is not specified anywhere by the SD specs.
On U-boot code, it states this is a valid command for eSD cards, but it
doesnt check if the card is an eSD before sending it. However, U-boot
code does not test for failure of the command and only continues
execution as if nothing had happened. It looks like Uboot takes
advantage of ignoring the illegal command to avoiding duplicated code.

This patch only adds a dumb return for the cmd43, allowing the code to
continue and not falling into an assert trap.