This patch changes the method used by the GDB stub to fetch a  byte from memory.
The GDB stub used 4 MEM.read_bytes to fetch a word from memory. This
executes an unaligned memory access to fetch each byte from the
requested address.  The problem with this approach is that unaligned
access are not supported by memory-mapped devices, so when GDB tries to
read a register from a memory-mapped device, we would get unpredictable
values.

Now, we perform only one aligned word-size memory read and select the
correct byte on the GDB stub.  This approach fix the previous
inconsistency and is more efficient.